+++ src/core/isAttached.js
@@ -1,8 +1,6 @@
 import jQuery from "../core.js";
 import documentElement from "../var/documentElement.js";
 
-import "../selector/contains.js"; // jQuery.contains
-
 var isAttached = function( elem ) {
 		return jQuery.contains( elem.ownerDocument, elem ) ||
 			elem.getRootNode( composed ) === elem.ownerDocument;
